requirements.txt
Flask==2.0.3
gunicorn==20.1.0
requests==2.26.0
beautifulsoup4==4.13.3
google-cloud-firestore==2.20.1
python-dotenv==1.1.0
urllib3==1.26.14
google-cloud-core==2.4.3
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9
soupsieve==2.6
typing_extensions==4.13.0
Werkzeug==2.0.3


